refactor: remove glassmorphism
