+++ packages/client/src/components/base/ErrorBoundary.tsx
