+++ packages/client/src/styles/common.ts
@@ -16,7 +16,7 @@ const customMediaQuery = (maxWidth: number) => `
 export const mediaQuery = {
   custom: customMediaQuery,
   mobile: customMediaQuery(500),
-  tablet: customMediaQuery(1024),
+  tablet: customMediaQuery(768),
   desktop: customMediaQuery(1440),
 } as const;
